% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDOLOptimizedCpGsCordBlood.R
\docType{data}
\name{IDOLOptimizedCpGsCordBlood}
\alias{IDOLOptimizedCpGsCordBlood}
\title{IDOL Optimized CpGs for umbilical cord blood DNA methylation 
    deconvolution}
\format{An object of class "character" of length 517.

        The format is:
        chr [1:517] "cg12603453" "cg24765783" "cg06975018" "cg19708055" ...}
\usage{
IDOLOptimizedCpGsCordBlood
}
\description{
This object is a vector of length 517 consisting of the names of the IDOL
    optimized CpGs.  These CpGs are used as the backbone for deconvolution
    and were selected because their methylation signature differs across the 
    six normal leukocyte subtypes and the nucleated red blood cells.
}
\examples{
# Do not run
# library (FlowSorted.Blood.EPIC)
# data("IDOLOptimizedCpGsCordBlood")
# Step 1: Load the reference library to extract the umbilical cord samples
# library(ExperimentHub)
# hub <- ExperimentHub()
# myfiles <- query(hub, "FlowSorted.CordBloodCombined.450k")
# FlowSorted.CordBloodCombined.450k <- myfiles[[1]]
# FlowSorted.CordBloodCombined.450k

# Step 2 separate the reference from the testing dataset if you want to run 
# examples for estimations for this function example

# RGsetTargets <- FlowSorted.CordBloodCombined.450k[,
# FlowSorted.CordBloodCombined.450k$CellType == "WBC"]
# sampleNames(RGsetTargets) <- paste(RGsetTargets$CellType,
#                               seq_len(dim(RGsetTargets)[2]), sep = "_")
# RGsetTargets

# Step 3: use your favorite package for deconvolution.
# Deconvolute a target data set consisting of 450K DNA methylation 
# data profiled in blood, using your prefered method.
# You can use our IDOL optimized DMR library for the Cord Blood,  This object
# contains a vector of length 517 consisting of the IDs of the IDOL optimized
# CpG probes.  These CpGs are used as the backbone for deconvolution and were
# selected because their methylation signature differs across the six normal 
# leukocyte subtypes plus the nucleated red blood cells.

# data (IDOLOptimizedCpGsCordBlood)
# head (IDOLOptimizedCpGsCordBlood)
# We recommend using Noob processMethod = "preprocessNoob" in minfi for the 
# target and reference datasets. 
# Cell types included are "CD8T", "CD4T", "NK", "Bcell", "Mono", "Gran", 
# "nRBC"
# To use the IDOL optimized list of CpGs (IDOLOptimizedCpGsCordBlood) use 
# estimateCellCounts2 from FlowSorted.Blood.EPIC. 
# Do not run with limited RAM the normalization step requires a big amount 
# of memory resources. Use the parameters as specified below for 
# reproducibility.
# 
# if (memory.limit()>8000){
#     countsEPIC<-estimateCellCounts2(RGsetTargets, 
#                                     compositeCellType = "Blood", 
#                                     processMethod = "preprocessNoob",
#                                     probeSelect = "IDOL", 
#                                     cellTypes = c("CD8T", "CD4T", "NK",  
#                                     "Bcell", "Mono", "Gran", "nRBC"), 
#                                     referencePlatform = 
#                                         "IlluminaHumanMethylationEPIC",
#                                     referenceset = 
#                                      "FlowSorted.CordBloodCombined.450k",
#                                     IDOLOptimizedCpGs =
#                                       IDOLOptimizedCpGsCordBlood, 
#                                     returnAll = FALSE)
#     
#     head(countsEPIC$counts)
# }
}
\references{
K Gervin, LA Salas et al. (2019) \emph{Systematic evaluation and 
validation of references and library selection methods for deconvolution of 
cord blood DNA methylation data}. (Under review).

LA Salas et al. (2018). \emph{An optimized library for 
reference-based deconvolution of whole-blood biospecimens assayed using the 
Illumina HumanMethylationEPIC BeadArray}. Genome Biology 19, 64. doi:
10.1186/s13059-018-1448-7.

DC Koestler et al. (2016). \emph{Improving cell mixture 
deconvolution by identifying optimal DNA methylation libraries (IDOL)}. 
BMC bioinformatics. 17, 120. doi: 10.1186/s12859-016-0943-7.
}
\keyword{datasets}
